+++ lib/Models/Terria.ts
@@ -201,7 +201,7 @@ export default class Terria {
     conversionServiceBaseUrl: "convert/",
     proj4ServiceBaseUrl: "proj4/",
     corsProxyBaseUrl: "proxy/",
-    proxyableDomainsUrl: "proxyabledomains/",
+    proxyableDomainsUrl: "proxyabledomains/", // deprecated, will be determined from serverconfig
     serverConfigUrl: "serverconfig/",
     shareUrl: "share",
     feedbackUrl: undefined,
@@ -930,33 +930,16 @@ export default class Terria {
     }
   }
 
-  initCorsProxy(config: any, serverConfig: any): Promise<void> {
-    // All the "proxyableDomains" bits here are due to a pre-serverConfig mechanism for whitelisting domains.
-    // We should deprecate it.s
-
-    // If a URL was specified in the config parameters to get the proxyable domains from, get them from that
-    var pdu = this.configParameters.proxyableDomainsUrl;
-    const proxyableDomainsPromise: Promise<JsonValue | void> = pdu
-      ? loadJson5(pdu)
-      : Promise.resolve();
-    return proxyableDomainsPromise.then((proxyableDomains: any | void) => {
-      if (proxyableDomains) {
-        // format of proxyableDomains JSON file slightly differs from serverConfig format.
-        proxyableDomains.allowProxyFor =
-          proxyableDomains.allowProxyFor || proxyableDomains.proxyableDomains;
-      }
-
-      // If there isn't anything there, check the server config
-      if (typeof serverConfig === "object") {
-        serverConfig = serverConfig.config; // if server config is unavailable, this remains undefined.
-      }
-
-      this.corsProxy.init(
-        proxyableDomains || serverConfig,
-        this.configParameters.corsProxyBaseUrl,
-        config.proxyDomains // fall back to local config
-      );
-    });
+  initCorsProxy(config: ConfigParameters, serverConfig: any): Promise<void> {
+    if (config.proxyableDomainsUrl) {
+      console.warn(i18next.t("models.terria.proxyableDomainsDeprecation"));
+    }
+    this.corsProxy.init(
+      serverConfig,
+      this.configParameters.corsProxyBaseUrl,
+      []
+    );
+    return Promise.resolve();
   }
 
   getUserProperty(key: string) {