+++ test/Models/TerriaSpec.ts
@@ -330,6 +330,17 @@ describe("Terria", function() {
   });
 
   it("initializes proxy with parameters from config file", function(done) {
+    jasmine.Ajax.install();
+    jasmine.Ajax.stubRequest(/.*(test\/init\/configProxy).*/).andReturn({
+      responseText: JSON.stringify(
+        require("../../wwwroot/test/init/configProxy.json")
+      )
+    });
+    jasmine.Ajax.stubRequest(/.*(serverconfig).*/).andReturn({
+      responseText: JSON.stringify(
+        require("../../wwwroot/test/init/serverconfig.json")
+      )
+    });
     terria
       .start({
         configUrl: "test/init/configProxy.json"